+++ metrictank/metrictank.go
@@ -22,13 +22,13 @@ func Init(metrictankUrl string) error {
 	return err
 }
 
-func ProxyDelete(orgId int64) *httputil.ReverseProxy {
+func ProxyDelete(orgId int) *httputil.ReverseProxy {
 	director := func(req *http.Request) {
 		req.URL.Scheme = MetrictankUrl.Scheme
 		req.URL.Host = MetrictankUrl.Host
 		req.URL.Path = util.JoinUrlFragments(MetrictankUrl.Path, "/metrics/delete")
 		req.Header.Del("X-Org-Id")
-		req.Header.Add("X-Org-Id", strconv.FormatInt(orgId, 10))
+		req.Header.Add("X-Org-Id", strconv.FormatInt(int64(orgId), 10))
 	}
 	return &httputil.ReverseProxy{Director: director}
 }
